# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a simple RESTful CRM Service API written in Go using fiber web framework, PostgreSQL database",
        "title": "CRM API",
        "contact": {
            "name": "Roihan Sori",
            "email": "roihansori12@gmail.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "basePath": "/v1",
    "paths": {
...
        "/attendances": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "List Attendances",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Attendances"
                ],
                "summary": "List Attendances",
                "responses": {
                    "200": {
                        "description": "Attendances retrieved",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handler.AttendanceResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "create a new Attendance with employee_id, location_id, absent_in, and absent_out",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Attendances"
                ],
                "summary": "Create a new Attendance",
                "parameters": [
                    {
                        "type": "string",
                        "example": "2024-07-11T08:00:00Z",
                        "name": "absent_in",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "2024-07-11T17:00:00Z",
                        "name": "absent_out",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "Admin",
                        "name": "created_by",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "example": 1,
                        "name": "employee_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "example": 1,
                        "name": "location_id",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Attendance created",
                        "schema": {
                            "$ref": "#/definitions/handler.AttendanceResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Data not found error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/attendances/{attendanceId}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "get an Attendance by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Attendances"
                ],
                "summary": "Get an Attendance",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Attendance ID",
                        "name": "attendanceId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Attendance retrieved",
                        "schema": {
                            "$ref": "#/definitions/handler.AttendanceResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Data not found error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "update an Attendance with employee_id, location_id, absent_in, and absent_out",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Attendances"
                ],
                "summary": "Update an Attendance",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Attendance ID",
                        "name": "attendanceId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "2024-07-11T08:00:00Z",
                        "name": "absent_in",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "2024-07-11T17:00:00Z",
                        "name": "absent_out",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "example": 1,
                        "name": "employee_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "example": 1,
                        "name": "location_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "Admin",
                        "name": "updated_by",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Attendance updated",
                        "schema": {
                            "$ref": "#/definitions/handler.AttendanceResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Data not found error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "delete an Attendance by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Attendances"
                ],
                "summary": "Delete an Attendance",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Attendance ID",
                        "name": "attendanceId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Attendance deleted",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Data not found error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.AttendanceResponse": {
            "type": "object",
            "properties": {
                "absent_in": {
                    "type": "string"
                },
                "absent_out": {
                    "type": "string"
                },
                "attendance_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "employee_id": {
                    "type": "integer"
                },
                "location_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "utils.ErrorResponse": {
            "type": "object",
            "properties": {
                "messages": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "Error message 1",
                        " Error message 2"
                    ]
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "utils.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "Success"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and the access token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}